# demart-api-contract.yaml
openapi: 3.0.3
info:
  title: DeMart API
  version: 1.0.0
  description: REST API contract for the DeMart e-commerce application.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request a password reset
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Password reset request processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForgotPasswordResponse'
        '400':
          description: Invalid email address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many password reset requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset a password using a reset token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or expired token, or invalid password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/locations/countries:
    get:
      tags: [Locations]
      summary: Get available countries
      operationId: getCountries
      responses:
        '200':
          description: Countries returned successfully
          content:
            application/json:
              schema:
                type: object
                required: [countries]
                properties:
                  countries:
                    type: array
                    items:
                      type: object
                      required: [country]
                      properties:
                        country:
                          type: string
        '500':
          description: Failed to load countries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/locations/regions:
    get:
      tags: [Locations]
      summary: Get regions for a country
      operationId: getRegions
      parameters:
        - name: country
          in: query
          required: true
          description: Country name.
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Regions returned successfully
          content:
            application/json:
              schema:
                type: object
                required: [regions]
                properties:
                  regions:
                    type: array
                    items:
                      type: object
                      required: [region]
                      properties:
                        region:
                          type: string
        '400':
          description: Country is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to load regions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/locations/cities:
    get:
      tags: [Locations]
      summary: Get cities for a country and region
      operationId: getCities
      parameters:
        - name: country
          in: query
          required: true
          description: Country name.
          schema:
            type: string
            minLength: 1
        - name: region
          in: query
          required: true
          description: Region name.
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Cities returned successfully
          content:
            application/json:
              schema:
                type: object
                required: [cities]
                properties:
                  cities:
                    type: array
                    items:
                      type: object
                      required: [id, city, timezone]
                      properties:
                        id:
                          type: integer
                        city:
                          type: string
                        timezone:
                          type: string
        '400':
          description: Country and region are required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to load cities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/stores:
    get:
      tags: [Stores]
      summary: Get stores for a location
      operationId: getStoresByLocation
      parameters:
        - name: locationId
          in: query
          required: true
          description: Location ID.
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Stores returned successfully
          content:
            application/json:
              schema:
                type: object
                required: [stores]
                properties:
                  stores:
                    type: array
                    items:
                      $ref: '#/components/schemas/Store'
        '400':
          description: Location ID is missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to load stores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ForgotPasswordResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        resetUrl:
          type: string
          format: uri
          nullable: true
          description: Development-only reset URL. Must not be exposed in production.
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
          minLength: 1
        newPassword:
          type: string
          format: password
          minLength: 8
          description: Must contain at least one uppercase letter, one lowercase letter, one number, and one special character.
...
    Store:
      type: object
      required:
        - id
        - location_id
        - name
        - address
        - timezone
      properties:
        id:
          type: integer
        location_id:
          type: integer
        name:
          type: string
        address:
          type: string
          nullable: true
        timezone:
          type: string
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Invalid request